pyproject.toml
[project]
name = "sttuhl"
version = "0.1.0"
description = ""
authors = [
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "numpy (>=2.3.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
